requirements.txt
Flask==2.3.3
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0